refactor(Transferable): Restructure interfaces to be opt-in transferable, simplifying the general ca
Previously developers always had to implement GenericWorkerMessage, whether or not they were using
Transferables. Now the interface is significantly simpler and you only need to pass the actual
object type as input and output. This simplificatrion is achieved by having methods on both sides to
select the transferable elements if present.

BREAKING CHANGE: fromWorker interface no longer uses GenericWorkerMessage, as input and output,
rather it now uses the plain Input and Output types. To indicate which parts of the Input message
are transferable, a new callback is added `selectTransferables`. Likewise in the worker side, a
selectTransferables method is added to the `DoWork` interface, and is required with the
`DoTransferableWork` interface. Implement this to indicate which elements of a message are
`Transferable`.